backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archery-backend"
version = "2.0.0"
description = "数据库管理平台 2.0 后端"
requires-python = ">=3.12"
dependencies = [
    # Web 框架
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    # ORM & 数据库
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",       # PostgreSQL async driver
    "psycopg2-binary>=2.9.9",# PostgreSQL sync driver (Alembic)
    # 数据验证
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    # 安全 & 认证
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",
    "pyotp>=2.9.0",           # Sprint 1: 2FA TOTP
    # Redis
    "redis[hiredis]>=5.0.1",
    # 异步任务
    "celery>=5.3.6",
    "redis>=5.0.1",
    "flower>=2.0.1",
    # SQL 解析（替代 goInception）
    "sqlglot>=23.0.0",
    # HTTP 客户端（内部 API 调用）
    "httpx>=0.27.0",
    # 对象存储
    "aioboto3>=12.3.0",
    # 数据库引擎驱动
    "aiomysql>=0.2.0",             # MySQL async
    "pymysql>=1.1.0",              # MySQL sync（备用）
    "pymongo>=4.6.0",              # MongoDB（替代 shell=True）
    "motor>=3.3.0",                # MongoDB async
    "aioredis>=2.0.1",             # Redis（引擎用）
    "clickhouse-driver>=0.2.7",    # ClickHouse
    "elasticsearch[async]>=8.12.0",# Elasticsearch
    "cassandra-driver>=3.29.0",    # Cassandra
    # OpenTelemetry 可观测性
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.44b0",
    "opentelemetry-exporter-prometheus>=0.44b0",
    # 结构化日志
    "structlog>=24.1.0",
    # AI 能力（可选）
    "anthropic>=0.20.0",
    "openai>=1.13.0",
    # 工具
    "python-dotenv>=1.0.1",
    "arrow>=1.3.0",               # 时间处理
    "jinja2>=3.1.3",              # 通知模板
]

[project.optional-dependencies]
oracle = [
    "cx_Oracle>=8.3.0",            # 需本地安装 Oracle Instant Client
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",              # TestClient
    "factory-boy>=3.3.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501", "B008", "N818"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
